requirements.txt
cachetools==5.5.2
certifi==2025.1.31
charset-normalizer==3.4.1
colorama==0.4.6
google-api-core==2.24.2
google-api-python-client==2.167.0
google-auth==2.39.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.1
googleapis-common-protos==1.70.0
greenlet==3.2.0
httplib2==0.22.0
idna==3.10
iniconfig==2.1.0
numpy==2.2.4
oauthlib==3.2.2
packaging==24.2
pandas==2.2.3
playwright==1.51.0
pluggy==1.5.0
proto-plus==1.26.1
protobuf==6.30.2
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycryptodome==3.22.0
pyee==12.1.1
pyparsing==3.2.3
pytest==8.3.5
python-dateutil==2.9.0.post0
pytz==2025.2
pysqlcipher3
PyPDF2==3.0.1
requests==2.32.3
requests-oauthlib==2.0.0
rsa==4.9.1
six==1.17.0
sqlalchemy
typing_extensions==4.13.2
tzdata==2025.2
uritemplate==4.1.1
urllib3==2.4.0
